requirements.txt
python-telegram-bot>=20.0
google-api-python-client>=2.0
google-auth>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.1
python-dotenv>=1.0
httpx>=0.24
